Ict2 45게시판 수정 대충 완료

+++ src/main/java/com/nutrimate/nutrimatebackend/controller/board/sport/SportBoardController.java
@@ -1,16 +1,15 @@
 package com.nutrimate.nutrimatebackend.controller.board.sport;
 
-import com.nutrimate.nutrimatebackend.model.FileUtils;
 import com.nutrimate.nutrimatebackend.model.board.sport.BookmarkDto;
 import com.nutrimate.nutrimatebackend.model.board.sport.LikeDto;
 import com.nutrimate.nutrimatebackend.model.board.sport.SportBoardDto;
 import com.nutrimate.nutrimatebackend.service.board.sport.SportBoardService;
 import jakarta.servlet.http.HttpServletRequest;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -19,6 +18,7 @@ import java.util.Map;
 
 @RestController
 @RequestMapping("/boards/sport")
+@Log4j2
 public class SportBoardController {
 	
 	@Autowired
@@ -26,17 +26,16 @@ public class SportBoardController {
 	
 	//글 생성
 	@PostMapping()
-	public SportBoardDto createBoard(SportBoardDto board, HttpServletRequest req, List<MultipartFile> files) {
-		String phisicalPath = "C:\\Temp\\upload";
-		StringBuffer fileNames = new StringBuffer();
-		try {
-			fileNames = FileUtils.upload(files, phisicalPath);
-			board.setMapImg(fileNames.toString());
-			
-		} catch (Exception e) { //파일용량 초과시
-			System.out.println("문제생김");
+	public Map<String, String> createBoard(@RequestBody SportBoardDto board, HttpServletRequest req) {
+		Map<String, String> map = new HashMap<>();
+		SportBoardDto createdBoard = sportBoardService.createBoard(board);
+		if (createdBoard == null) {
+			map.put("message", "글 작성에 실패했습니다");
+		} else {
+			map.put("message", "글 작성에 성공했습니다");
+			map.put("boardId", String.valueOf(createdBoard.getBoardId())); // boardId 추가
 		}
-		return sportBoardService.createBoard(board);
+		return map;
 	}
 	
 	//글 조회(상세)
@@ -62,45 +61,49 @@ public class SportBoardController {
 		return response.isEmpty() ? new ResponseEntity<>(HttpStatus.NO_CONTENT) : new ResponseEntity<>(response, HttpStatus.OK);
 	}
 	
-	
 	//글 목록 조회(전체)
 	@GetMapping()
 	public ResponseEntity<Map<String, Object>> getAllBoards(
-			@RequestParam(defaultValue = "1") int pageNum,
-			@RequestParam(required = false) String searchUser,
-			@RequestParam(required = false) String searchTitle,
-			@RequestParam(required = false) String searchTag) {
+			@RequestParam(name = "pageNum", defaultValue = "1") int pageNum,
+			@RequestParam(name = "searchUser", required = false) String searchUser,
+			@RequestParam(name = "searchTitle", required = false) String searchTitle,
+			@RequestParam(name = "searchContent", required = false) String searchContent,
+			@RequestParam(name = "searchTag", required = false) String searchTag) {
 		
-		Map<String, Object> boards = sportBoardService.getBoards(pageNum, searchUser, searchTitle, searchTag);
+		Map<String, Object> boards = sportBoardService.getBoards(pageNum, searchUser, searchTitle, searchContent, searchTag);
 		return new ResponseEntity<>(boards, HttpStatus.OK);
 	}
 	
 	//글 수정
 	@PutMapping("/{id}")
-	public SportBoardDto updateBoard(@PathVariable("id") int id, SportBoardDto board, HttpServletRequest req, List<MultipartFile> files) {
+	public Map<String, String> updateBoard(@PathVariable("id") int id, @RequestBody SportBoardDto board, HttpServletRequest req) {
 		board.setBoardId(id);
-		String phisicalPath = "C:\\Temp\\upload";
-		StringBuffer fileNames = new StringBuffer();
-		if (files != null) {
-			try {
-				
-				SportBoardDto titledto = sportBoardService.getBoard(id);
-				System.out.println(titledto);
-				StringBuffer titledto_ = new StringBuffer(titledto.getMapImg());
-				fileNames = FileUtils.upload(files, phisicalPath);
-				FileUtils.deletes(titledto_, phisicalPath, ",");
-				board.setMapImg(fileNames.toString());
-			} catch (Exception e) {//파일용량 초과시
-			
-			}
+		
+		Map<String, String> map = new HashMap<>();
+		SportBoardDto updatedBoard = sportBoardService.updateBoard(board);
+		
+		if (updatedBoard == null) {
+			map.put("message", "글 수정에 실패했습니다");
+			return map;
+		} else {
+			map.put("message", "글 수정에 성공했습니다");
+			return map;
 		}
-		return sportBoardService.updateBoard(board);
 	}
 	
 	//글 삭제
 	@DeleteMapping("/{id}")
-	public void deleteBoard(@PathVariable("id") int id) {
-		sportBoardService.deleteBoard(id);
+	public Map<String, String> deleteBoard(@PathVariable("id") int id) {
+		Map<String, String> map = new HashMap<>();
+		
+		try {
+			sportBoardService.deleteBoard(id);
+			map.put("message", "글 삭제에 성공했습니다");
+		} catch (Exception e) {
+			map.put("message", "글 삭제에 실패했습니다");
+		}
+		
+		return map;
 	}
 	
 	//좋아요 수 조회
@@ -109,32 +112,95 @@ public class SportBoardController {
 		return sportBoardService.countLikes(boardId);
 	}
 	
+	// 유저가 좋아요 누른지 확인 (완료)
+	// 입력 데이터 : userId, boardId
+	// 출력 데이터 : message, boardId, userId
+	@PostMapping("/like/check")
+	public ResponseEntity<Map<String, Object>> checkUserLike(@RequestBody LikeDto likeDto) {
+		int checkUserLike = sportBoardService.checkUserLike(likeDto);
+		Map<String, Object> jsonResponse = new HashMap<>();
+		if (checkUserLike == 0) {
+			jsonResponse.put("message", "좋아요를 안눌렀어요");
+		} else {
+			jsonResponse.put("message", "좋아요를 이미 눌렀어요");
+		}
+		jsonResponse.put("boardId", likeDto.getBoardId());
+		jsonResponse.put("userId", likeDto.getUserId());
+		return ResponseEntity.ok(jsonResponse);
+	}
+	
 	//좋아요 생성
 	@PostMapping("/{boardId}/likes")
-	public ResponseEntity<Void> insertLike(@PathVariable String boardId, @RequestBody LikeDto likeDto) {
+	public ResponseEntity<Map<String, String>> insertLike(@PathVariable String boardId, @RequestBody LikeDto likeDto) {
+		Map<String, String> response = new HashMap<>();
 		likeDto.setBoardId(boardId);
+		boolean isLiked = sportBoardService.isLiked(likeDto);
+		
+		if (isLiked) {
+			response.put("message", "이미 좋아요를 눌렀습니다");
+			return new ResponseEntity<>(response, HttpStatus.OK); // 상태 코드를 OK로 변경
+		}
+		
 		boolean isCreated = sportBoardService.insertLike(likeDto);
 		
 		if (isCreated) {
-			return new ResponseEntity<>(HttpStatus.CREATED);
+			response.put("message", "좋아요 추가 성공");
+			return new ResponseEntity<>(response, HttpStatus.CREATED);
 		} else {
-			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			response.put("message", "좋아요 추가 실패");
+			return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
 	
 	//좋아요 삭제
 	@DeleteMapping("/{boardId}/likes")
-	public ResponseEntity<Void> deleteLike(@PathVariable String boardId, @RequestBody LikeDto likeDto) {
+	public ResponseEntity<Map<String, String>> deleteLike(@PathVariable String boardId, @RequestBody LikeDto likeDto) {
+		Map<String, String> response = new HashMap<>();
 		likeDto.setBoardId(boardId);
+		int countLikes = sportBoardService.countLikes(boardId);
+		log.info("boardId : " + boardId);
+		log.info("countLikes : " + countLikes);
+		
+		if (countLikes == 0) {
+			response.put("message", "좋아요를 누르지 않았습니다");
+			return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+		}
+		
 		int affectedRows = sportBoardService.deleteLike(likeDto);
 		
 		if (affectedRows > 0) {
-			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+			response.put("message", "좋아요 삭제 성공");
+			return new ResponseEntity<>(response, HttpStatus.OK);
 		} else {
-			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			response.put("message", "좋아요 삭제 실패");
+			return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
+	}	
+	/*
+	@DeleteMapping("/{boardId}/likes")
+	public ResponseEntity<Map<String, String>> deleteLike(@PathVariable String boardId, @RequestBody LikeDto likeDto) {
+	    Map<String, String> response = new HashMap<>();
+	    likeDto.setBoardId(boardId);
+
+	    boolean isLikedByUser = sportBoardService.isLiked(likeDto);
+	    
+	    if (!isLikedByUser) {
+	        response.put("message", "좋아요를 누르지 않았습니다");
+	        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+	    }
+
+	    int affectedRows = sportBoardService.deleteLike(likeDto);
+	    
+	    if (affectedRows > 0) {
+	        response.put("message", "좋아요 삭제 성공");
+	        return new ResponseEntity<>(response, HttpStatus.OK);
+	    } else {
+	        response.put("message", "좋아요 삭제 실패");
+	        return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+	    }
 	}
-   
+  	*/
+	
     /*
     //이전글 조회
     @GetMapping("/{id}/prev")
@@ -161,37 +227,55 @@ public class SportBoardController {
 	
 	//북마크 생성
 	@PostMapping("/{boardId}/bookmark")
-	public ResponseEntity<Void> insertBookmark(@PathVariable String boardId, @RequestBody BookmarkDto bookmarkDto) {
+	public ResponseEntity<Map<String, String>> insertBookmark(@PathVariable String boardId, @RequestBody BookmarkDto bookmarkDto) {
+		Map<String, String> response = new HashMap<>();
 		bookmarkDto.setBoardId(boardId);
+		boolean isBookmarked = sportBoardService.isBookmarked(bookmarkDto);
+		
+		if (isBookmarked) {
+			response.put("message", "이미 북마크를 추가했습니다");
+			return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+		}
+		
 		boolean isCreated = sportBoardService.insertBookmark(bookmarkDto);
 		
 		if (isCreated) {
-			return new ResponseEntity<>(HttpStatus.CREATED);
+			response.put("message", "북마크 추가 성공");
+			return new ResponseEntity<>(response, HttpStatus.CREATED);
 		} else {
-			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			response.put("message", "북마크 추가 실패");
+			return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
 	
 	//북마크 삭제
 	@DeleteMapping("/{boardId}/bookmark")
-	public ResponseEntity<Void> deleteBookmark(@PathVariable String boardId, @RequestBody BookmarkDto bookmarkDto) {
+	public ResponseEntity<Map<String, String>> deleteBookmark(@PathVariable String boardId, @RequestBody BookmarkDto bookmarkDto) {
+		Map<String, String> response = new HashMap<>();
 		bookmarkDto.setBoardId(boardId);
+		boolean isBookmarked = sportBoardService.isBookmarked(bookmarkDto);
+		
+		if (!isBookmarked) {
+			response.put("message", "북마크를 추가하지 않았습니다");
+			return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+		}
+		
 		int affectedRows = sportBoardService.deleteBookmark(bookmarkDto);
 		
 		if (affectedRows > 0) {
-			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+			response.put("message", "북마크 삭제 성공");
+			return new ResponseEntity<>(response, HttpStatus.OK);
 		} else {
-			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			response.put("message", "북마크 삭제 실패");
+			return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
 	
-	/** 해시태그 **/
-	// 해당 글의 해시태그 가져오기 (완료)
-	// 입력 데이터 : boardId
-	// 출력 데이터 : tagName or message(해시태그가 비어있거나 없는 게시글번호)
-	@GetMapping("/hashtag")
-	public ResponseEntity<List<Map<String, Object>>> findHashtagsByBoardId(@RequestBody SportBoardDto board) {
-		List<SportBoardDto> tagNames = sportBoardService.findHashtagsByBoardId(board);
+	//해당 글의 해시태그 가져오기
+	//입력 데이터:boardId / 출력 데이터:tagName or message(해시태그가 비어있거나 없는 게시글번호)
+	@GetMapping("/{boardId}/hashtag")
+	public ResponseEntity<List<Map<String, Object>>> findHashtagsByBoardId(@PathVariable String boardId) {
+		List<SportBoardDto> tagNames = sportBoardService.findHashtagsByBoardId(boardId);
 		List<Map<String, Object>> tagNameList = new ArrayList<>();
 		for (SportBoardDto tagName : tagNames) {
 			Map<String, Object> tagNameBoard = new HashMap<>();
@@ -200,7 +284,7 @@ public class SportBoardController {
 		}
 		if (tagNameList.isEmpty()) {
 			Map<String, Object> tagNameBoard = new HashMap<>();
-			tagNameBoard.put("message", "해시태그가 없어요");
+			tagNameBoard.put("message", "해시태그가 없습니다");
 			tagNameList.add(tagNameBoard);
 			return new ResponseEntity<>(tagNameList, HttpStatus.OK);
 		}
@@ -210,24 +294,24 @@ public class SportBoardController {
 	// 해시태그로 글 검색 (완료)
 	// 입력 데이터 : tagName
 	// 출력 데이터 : BoardList or message(해당 해시태그로 검색된 글이 없음)
-	/*
 	@GetMapping("/hashtag/search")
-	public ResponseEntity<List<Map<String, Object>>> findBoardsByTagName(@RequestBody BoardDto board) {
-	    List<BoardDto> findBoardsByTagNames = boardService.findBoardsByTagName(board);
-	    int boardId = board.getBoardId();
-	    List<Map<String, Object>> BoardsByTagNameList = new ArrayList<>();
-	    for (BoardDto findBoardsByTagName : findBoardsByTagNames) {
-	      Map<String, Object> BoardList = new HashMap<>();
-	      BoardList.put("BoardList", findBoardsByTagName.getBoardId());
-	      BoardsByTagNameList.add(BoardList);
-	    }
-	    if (BoardsByTagNameList.isEmpty()) {
-	      Map<String, Object> tagNameFeed = new HashMap<>();
-	      tagNameFeed.put("message", "해당 해시태그로 검색된 글이 없어요");
-	      BoardsByTagNameList.add(tagNameFeed);
-	      return new ResponseEntity<>(BoardsByTagNameList, HttpStatus.OK);
-	    }
-	    return new ResponseEntity<>(BoardsByTagNameList, HttpStatus.OK);
+	public ResponseEntity<List<Map<String, Object>>> findBoardsByTagName(
+			@RequestBody SportBoardDto board) {
+		List<SportBoardDto> findBoardsByTagNames = sportBoardService.findBoardsByTagName(board);
+		int boardId = board.getBoardId();
+		List<Map<String, Object>> BoardsByTagNameList = new ArrayList<>();
+		for (SportBoardDto findBoardsByTagName : findBoardsByTagNames) {
+			Map<String, Object> BoardList = new HashMap<>();
+			BoardList.put("BoardList", findBoardsByTagName.getBoardId());
+			BoardsByTagNameList.add(BoardList);
+		}
+		if (BoardsByTagNameList.isEmpty()) {
+			Map<String, Object> tagNameBoard = new HashMap<>();
+			tagNameBoard.put("message", "해당 해시태그로 검색된 글이 없어요");
+			BoardsByTagNameList.add(tagNameBoard);
+			return new ResponseEntity<>(BoardsByTagNameList, HttpStatus.OK);
+		}
+		return new ResponseEntity<>(BoardsByTagNameList, HttpStatus.OK);
 	}
-	*/
+	
 }
